Move SignalProducer APIs to a new source file.
